# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Test task for messaggio",
        "title": "Messaggio REST API",
        "contact": {
            "name": "Siarhei Vasileuski",
            "url": "https://t.me/kataomione",
            "email": "sergej.vasilewsckij@yandex.ru"
        },
        "version": "1.0"
    },
    "paths": {
        "/msg": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Receive new message",
                "parameters": [
                    {
                        "description": "Message content",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comments"
                ],
                "summary": "Get messages statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Statistics"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreatedResponse": {
            "type": "object",
            "properties": {
                "messageID": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Request": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "models.Statistics": {
            "type": "object",
            "properties": {
                "allMessages": {
                    "type": "integer"
                },
                "lastProcessed": {
                    "type": "string"
                },
                "processedMessages": {
                    "type": "integer"
                }
            }
        }
    }
}